move polynomiallr to examples/mlperf
